Move all breadboard-web config goop into one file.
