+++ packages/breadboard-web/src/index.ts
@@ -6,42 +6,10 @@
 
 import * as BreadboardUI from "@google-labs/breadboard-ui";
 import {
-  type HarnessProxyConfig,
-  type HarnessRemoteConfig,
   createHarness,
   HarnessRunResult,
 } from "@google-labs/breadboard/harness";
-import { asRuntimeKit } from "@google-labs/breadboard";
-import Starter from "@google-labs/llm-starter";
-import PaLMKit from "@google-labs/palm-kit";
-import NodeNurseryWeb from "@google-labs/node-nursery-web";
-import Core from "@google-labs/core-kit";
-import Pinecone from "@google-labs/pinecone-kit";
-import JSONKit from "@google-labs/json-kit";
-
-const PROXY_NODES = [
-  "palm-generateText",
-  "palm-embedText",
-  "secrets",
-  "fetch",
-  // TODO: These are only meaningful when proxying to main thread,
-  //       not anywhere else. Need to figure out what to do here.
-  // "credentials",
-  // "driveList",
-];
-
-const WORKER_URL =
-  import.meta.env.MODE === "development" ? "/src/worker.ts" : "/worker.js";
-
-const HARNESS_SWITCH_KEY = "bb-harness";
-
-const PROXY_SERVER_HARNESS_VALUE = "proxy-server";
-const WORKER_HARNESS_VALUE = "worker";
-
-const PROXY_SERVER_URL = import.meta.env.VITE_PROXY_SERVER_URL;
-const DEFAULT_HARNESS = PROXY_SERVER_URL
-  ? PROXY_SERVER_HARNESS_VALUE
-  : WORKER_HARNESS_VALUE;
+import { createHarnessConfig } from "./config";
 
 type PauserCallback = (paused: boolean) => void;
 class Pauser extends EventTarget {
@@ -119,7 +87,7 @@ export class Main {
         const startEvent = evt as BreadboardUI.StartEvent;
         this.setActiveBreadboard(startEvent.url);
 
-        const harness = this.#getHarness(startEvent.url);
+        const harness = createHarness(createHarnessConfig(startEvent.url));
 
         this.#ui.load(await harness.load());
 
@@ -285,36 +253,5 @@ export class Main {
         this.#hasActiveBoard = false;
         break;
     }
-  }
-
-  #getHarness(url: string) {
-    const harness =
-      globalThis.localStorage.getItem(HARNESS_SWITCH_KEY) ?? DEFAULT_HARNESS;
-
-    const kits = [
-      Starter,
-      Core,
-      Pinecone,
-      PaLMKit,
-      NodeNurseryWeb,
-      JSONKit,
-    ].map((kitConstructor) => asRuntimeKit(kitConstructor));
-
-    const proxy: HarnessProxyConfig[] = [];
-    if (harness === PROXY_SERVER_HARNESS_VALUE) {
-      proxy.push({
-        location: "http",
-        url: PROXY_SERVER_URL,
-        nodes: PROXY_NODES,
-      });
-    } else if (harness === WORKER_HARNESS_VALUE) {
-      proxy.push({ location: "main", nodes: PROXY_NODES });
-    }
-    const remote: HarnessRemoteConfig = harness === WORKER_HARNESS_VALUE && {
-      type: "worker",
-      url: WORKER_URL,
-    };
-    const diagnostics = true;
-    return createHarness({ url, kits, remote, proxy, diagnostics });
   }
 }
